+++ application_sdk/activities/metadata_extraction/sql.py
@@ -143,8 +143,8 @@ class BaseSQLMetadataExtractionActivities(ActivitiesInterface):
         self._state[workflow_id].handler = handler
 
         if "credential_guid" in workflow_args:
-            credentials = SecretStoreInput.extract_credentials(
-                workflow_args["credential_guid"]
+            credentials = await SecretStoreInput.fetch_secret(
+                secret_key=workflow_args["credential_guid"]
             )
             await sql_client.load(credentials)